Stash issue happened in GRNService file

+++ src/main/java/com/pos/hyper/model/grn/GRNService.java
@@ -3,6 +3,7 @@ package com.pos.hyper.model.grn;
 import com.pos.hyper.DTO.GRNDto;
 import com.pos.hyper.exception.CustomExceptionHandler;
 import com.pos.hyper.repository.GRNRepository;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,36 +21,45 @@ public class GRNService {
     }
 
 
-    public GRNDto createGRN(GRNDto grnDto) {
+    public ResponseEntity<?> createGRN(GRNDto grnDto) {
         GRN grn = grnMapper.toGRN(grnDto);
         grn.setId(null);
         grn.setTotal(0.0);
         grn = grnRepository.save(grn);
-        return grnMapper.toGRNDto(grn);
+        return ResponseEntity.ok(grnMapper.toGRNDto(grn));
     }
     @Transactional
-    public GRNDto updateGRN(Integer id, GRNDto grnDto) {
-        GRN grn = grnRepository.findById(id)
-                .orElseThrow(() -> customExceptionHandler.handleNotFoundException("GRN with id " + id + " not found"));
+    public ResponseEntity<?> updateGRN(Integer id, GRNDto grnDto) {
+        GRN grn = grnRepository.findById(id).orElse(null);
+        if(grn == null) {
+            return customExceptionHandler.notFoundException("GRN with id " + id + " not found");
+        }
         double total = grn.getTotal();
         grn = grnMapper.toGRN(grnDto, grn);
         grn.setTotal(total);
         grn = grnRepository.save(grn);
-        return grnMapper.toGRNDto(grn);
+        return ResponseEntity.ok(grnMapper.toGRNDto(grn));
     }
-    public List<GRNDto> getAllGRNs() {
-        List<GRN> inventories = grnRepository.findAll();
-        return inventories.stream().map(grnMapper::toGRNDto).toList();
+    public ResponseEntity<?> getAllGRNs() {
+        List<GRN> GRNs = grnRepository.findAll();
+        return ResponseEntity.ok(GRNs.stream().map(grnMapper::toGRNDto).toList());
+
     }
-    public GRNDto getGRNById(Integer id) {
-        GRN grn = grnRepository.findById(id)
-                .orElseThrow(() -> customExceptionHandler.handleNotFoundException("GRN with id " + id + " not found"));
-        return grnMapper.toGRNDto(grn);
+    public ResponseEntity<?> getGRNById(Integer id) {
+        GRN grn = grnRepository.findById(id).orElse(null);
+        if (grn == null) {
+            return customExceptionHandler.notFoundException("GRN with id " + id + " not found");
+        }
+        return ResponseEntity.ok(grnMapper.toGRNDto(grn));
     }
-    public void deleteGRN(Integer id) {
-        GRN grn = grnRepository.findById(id)
-                .orElseThrow(() -> customExceptionHandler.handleNotFoundException("GRN with id " + id + " not found"));
+    public ResponseEntity<?> deleteGRN(Integer id) {
+        GRN grn = grnRepository.findById(id).orElse(null);
+        if(grn == null) {
+            return customExceptionHandler.notFoundException("GRN with id " + id + " not found");
+        }
         grnRepository.delete(grn);
+        return ResponseEntity.ok().build();
+
     }
